Reveal the mysterious behind the self assigning line
https://stackoverflow.com/a/47945948

Basically, we either should have something in our class that references
to the slot which the signal is connected or we need to keep a reference
to our class instances in order to not let them garbage collected.

So for the first case, if we replace the self assigning and connection
lines with:
self.a = self.get_possible_moves
board.mouse_clicked.connect(self.get_possible_moves)

our code would work exactly the same since we are increasing the
reference count to our method.

For the second case, we can hold all the pieces we created so we don't
even let them destroyed as I'm doing in this commit.

Lastly, the oldest version was also working with lambda's because we were
creating a closure, unknowingly, which let us keep a reference to our
objects even if they were deleted.

So writing the connection line with lambda's would always work.